Change `initial_magnetization` to `initial_magnetic_moments`
The protocol builder interface in `aiida-quantumespresso` change the
`initial_magnetization` input to `initial_magnetic_moments` because the
latter is more universal and easier to reason about. On top of that,
this way one does not have to have knowledge about the pseudo potentials
because the number of electrons in it determine the magnetization for a
given magnetic moment. This was causing a catch 22 in the input protocol
builder method.